// app/types/education.type.ts
export interface Education {
    id: string;
    collegeName: string;
    from: string;
    to: string;
    college: string;
    course: string;
    grade: string;
    candidate: string[];
    university: string;
    degree: string;
    subject: string;
    institution: string;
    university1: string;
    cpga: string;
    passOut: string;
}